src/etl/requirements.txt
requests
pandas
gspread
google-auth
numpy
scikit-learn